Notification Recovery and Log
This is a project that demonstrates the use of the notification listener in android to monitor notifications.
The project also explores the use of dependency injection using Hilt and the Work manager library to perform one time tasks in the background.
The data from the notifications is stored in a Room Database